test: relocate TestGethWriterBasic; delete Erigon-specific tests
Preparing the Erigon-removal PR by pulling out tests whose types will
disappear in the next commits. Three changes:

1. Move TestGethWriterBasic from generator/writer_erigon_test.go into a
   new generator/writer_geth_test.go. This test is the only place that
   exercises NewGethWriter / WriteAccount / WriteStorage / Flush /
   Stats in isolation — worth keeping.
2. Delete generator/writer_erigon_test.go (the two ErigonWriter tests
   that remain after the move would fail to compile once the writer
   goes).
3. Delete e2e_test.go's TestEndToEndErigonFormat and
   TestBothFormatsProduceSameStateRoot. The latter's whole purpose was
   cross-backend equivalence; with one backend remaining it has no
   meaning.

go test ./... still passes. No production code changes.
